pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sergei Mikurov <454004@mail.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
